Added image to be set on the repository level.  So it does not look so confusing when testing.  Change moment to dayJs and install source map explorer.
