+++ README.md
@@ -1,3 +1,7 @@
+## Demo Site
+   
+   [Click Here](http://d20qof6e8b8d9s.cloudfront.net/)
+
 1. ### Docker Setup
 
    [Click Here](https://github.com/TommyC5/test-full-stack/tree/tc_sf_fullstack_test/docker)
@@ -9,3 +13,5 @@
 3. ### Site Setup
 
    [Click Here](https://github.com/TommyC5/test-full-stack/tree/tc_sf_fullstack_test/site)
+   
+
